refactor: move settings to own file
This adds a partial class; I am not a fan of partial classes,
but in this case the one-click refactoring power wins over
my desire to remove complexity.
